==> 21_ReturningUsers.ipynb <==
# Returning User Modeling

IDEAS
+ more businesss meaning?
+ give data of user, predict/model on what?
  + when will the user return? implication on returning purchase
    + time prediction 
  + what would be the next review?
    + NLP sentence generation
  + what would be the next rating? 
    + numeric prediction
  + what product would the user review on? i.e. recommendation?
    + recommendation 

Predict number of review of a Product category?

##
